Restructure admin panel with side nav
