+++ src/app/pages/admin/admin.component.html
@@ -2,49 +2,16 @@
   <mat-sidenav mode="side" opened class="admin-sidenav">
     <mat-toolbar color="primary">ניהול</mat-toolbar>
     <mat-nav-list>
-      <a mat-list-item>רשומות</a>
+      <a mat-list-item routerLink="dashboard" routerLinkActive="active">דשבורד</a>
+      <a mat-list-item routerLink="submissions" routerLinkActive="active">רשומות</a>
     </mat-nav-list>
   </mat-sidenav>
   <mat-sidenav-content>
-    <mat-toolbar color="primary">
+    <mat-toolbar color="primary" class="toolbar">
       <span>לוח ניהול</span>
-      <span class="spacer"></span>
-      <button mat-icon-button (click)="openAddDialog()">
-        <mat-icon>add</mat-icon>
-      </button>
     </mat-toolbar>
     <div class="content">
-      <table mat-table [dataSource]="submissions" class="mat-elevation-z1" multiTemplateDataRows>
-        <ng-container matColumnDef="user">
-          <th mat-header-cell *matHeaderCellDef>משתמש</th>
-          <td mat-cell *matCellDef="let s">{{ s.userIdentity }}</td>
-        </ng-container>
-        <ng-container matColumnDef="formType">
-          <th mat-header-cell *matHeaderCellDef>טופס</th>
-          <td mat-cell *matCellDef="let s">{{ s.formName }}</td>
-        </ng-container>
-        <ng-container matColumnDef="timestamp">
-          <th mat-header-cell *matHeaderCellDef>תאריך</th>
-          <td mat-cell *matCellDef="let s">{{ s.timestamp | date:'short' }}</td>
-        </ng-container>
-        <ng-container matColumnDef="status">
-          <th mat-header-cell *matHeaderCellDef>סטטוס</th>
-          <td mat-cell *matCellDef="let s">{{ s.status }}</td>
-        </ng-container>
-        <ng-container matColumnDef="actions">
-          <th mat-header-cell *matHeaderCellDef>פעולות</th>
-          <td mat-cell *matCellDef="let s">
-            <button mat-icon-button color="primary" (click)="markProcessed(s)">
-              <mat-icon>check</mat-icon>
-            </button>
-            <button mat-icon-button color="warn" (click)="deleteSubmission(s.id!)">
-              <mat-icon>delete</mat-icon>
-            </button>
-          </td>
-        </ng-container>
-        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-        <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-      </table>
+      <router-outlet></router-outlet>
     </div>
   </mat-sidenav-content>
 </mat-sidenav-container>
