+++ src/app/pages/admin/admin.component.scss
@@ -4,9 +4,11 @@
 .admin-sidenav {
   width: 200px;
 }
+.toolbar {
+  position: sticky;
+  top: 0;
+  z-index: 1;
+}
 .content {
   padding: 1rem;
-}
-.spacer {
-  flex: 1 1 auto;
 }
