+++ src/app/pages/admin/dashboard/dashboard.component.html
@@ -1,0 +1,6 @@
+<mat-card>
+  <mat-card-title>דשבורד</mat-card-title>
+  <mat-card-content>
+    <p>בחר פריט בתפריט כדי להתחיל.</p>
+  </mat-card-content>
+</mat-card>
