+++ src/app/pages/admin/dashboard/dashboard.component.scss
@@ -1,0 +1,1 @@
+/* Empty placeholder */
