+++ src/app/pages/admin/submissions.component.html
@@ -1,0 +1,39 @@
+<div class="submissions-header">
+  <button mat-flat-button color="primary" (click)="openAddDialog()">
+    <mat-icon>add</mat-icon>
+    הוסף
+  </button>
+</div>
+<div class="content">
+  <table mat-table [dataSource]="submissions" class="mat-elevation-z1" multiTemplateDataRows>
+        <ng-container matColumnDef="user">
+          <th mat-header-cell *matHeaderCellDef>משתמש</th>
+          <td mat-cell *matCellDef="let s">{{ s.userIdentity }}</td>
+        </ng-container>
+        <ng-container matColumnDef="formType">
+          <th mat-header-cell *matHeaderCellDef>טופס</th>
+          <td mat-cell *matCellDef="let s">{{ s.formName }}</td>
+        </ng-container>
+        <ng-container matColumnDef="timestamp">
+          <th mat-header-cell *matHeaderCellDef>תאריך</th>
+          <td mat-cell *matCellDef="let s">{{ s.timestamp | date:'short' }}</td>
+        </ng-container>
+        <ng-container matColumnDef="status">
+          <th mat-header-cell *matHeaderCellDef>סטטוס</th>
+          <td mat-cell *matCellDef="let s">{{ s.status }}</td>
+        </ng-container>
+        <ng-container matColumnDef="actions">
+          <th mat-header-cell *matHeaderCellDef>פעולות</th>
+          <td mat-cell *matCellDef="let s">
+            <button mat-icon-button color="primary" (click)="markProcessed(s)">
+              <mat-icon>check</mat-icon>
+            </button>
+            <button mat-icon-button color="warn" (click)="deleteSubmission(s.id!)">
+              <mat-icon>delete</mat-icon>
+            </button>
+          </td>
+        </ng-container>
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+  </table>
+</div>
