+++ src/app/pages/admin/submissions.component.scss
@@ -1,0 +1,8 @@
+.submissions-header {
+  display: flex;
+  justify-content: flex-end;
+  margin-bottom: 1rem;
+}
+.content {
+  padding: 0 1rem;
+}
